+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Changelog
 - Fix broken test badge, see https://github.com/hpsim/OBR/pull/178
 - Fix decomposePar issues with non ESI versions of OF, see https://github.com/hpsim/OBR/issues/185
 - Automatically create tempory 0 folder before decomposePar, see https://github.com/hpsim/OBR/pull/186
+- Fix logfile location for shell commands, see https://github.com/hpsim/OBR/pull/184
+
 
 
 0.2.0 (2023-09-14)